backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
google-cloud-videointelligence==2.12.0
google-cloud-storage==2.14.0
google-generativeai==0.3.2
opencv-python-headless==4.9.0.80
ffmpeg-python==0.2.0
imagehash==4.3.1
rapidfuzz==3.6.1
pillow==10.2.0
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1